Use floats in commands where possible

+++ src/platz1de/EasyEdit/command/defaults/CountCommand.php
@@ -25,7 +25,7 @@ class CountCommand extends EasyEditCommand
 	public function process(Player $player, array $args): void
 	{
 		if (isset($args[0])) {
-			$selection = Sphere::aroundPoint($player->getName(), $player->getWorld()->getFolderName(), $player->getPosition(), (int) $args[0]);
+			$selection = Sphere::aroundPoint($player->getName(), $player->getWorld()->getFolderName(), $player->getPosition(), (float) $args[0]);
 		} else {
 			try {
 				$selection = SelectionManager::getFromPlayer($player->getName());
